split up vtcomposite-param-validation.test.js by function
